requirements.txt
torch==2.4.1
torchvision==0.19.1
torch-summary==1.4.5
numpy==1.26.4
pandas==2.1.4
matplotlib==3.7.5
scikit-learn==1.5.0
tqdm==4.66.4
tabulate==0.9.0
Pillow==10.2.0
IPython==8.25.0
